frontier/requirements.txt
matplotlib
numpy
pandas
scipy
scikit-learn
tensorflow-gpu
pyomo
xlrd